@@ -1,7 +1,7 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
     <PropertyGroup>
-        <TargetFramework>net7.0</TargetFramework>
+        <TargetFramework>net8.0</TargetFramework>
         <ImplicitUsings>disable</ImplicitUsings>
         <Nullable>enable</Nullable>
     </PropertyGroup>
@@ -21,23 +21,23 @@
         <PackageReference Include="EntityFrameworkCore.Exceptions.PostgreSQL" Version="6.0.3" />
         <PackageReference Include="EntityFrameworkCore.Exceptions.Sqlite" Version="6.0.3" />
         <PackageReference Include="EntityFrameworkCore.Exceptions.SqlServer" Version="6.0.3.2" />
-        <PackageReference Include="MassTransit" Version="8.1.1" />
-        <PackageReference Include="MassTransit.Azure.ServiceBus.Core" Version="8.1.1" />
-        <PackageReference Include="MassTransit.Kafka" Version="8.1.1" />
-        <PackageReference Include="MassTransit.RabbitMQ" Version="8.1.1" />
+        <PackageReference Include="MassTransit" Version="8.1.2" />
+        <PackageReference Include="MassTransit.Azure.ServiceBus.Core" Version="8.1.2" />
+        <PackageReference Include="MassTransit.Kafka" Version="8.1.2" />
+        <PackageReference Include="MassTransit.RabbitMQ" Version="8.1.2" />
         <PackageReference Include="Microsoft.AspNetCore.Http.Abstractions" Version="2.2.0" />
-        <PackageReference Include="Microsoft.Extensions.Configuration.Abstractions" Version="7.0.0" />
-        <PackageReference Include="Microsoft.Extensions.DependencyInjection.Abstractions" Version="7.0.0" />
+        <PackageReference Include="Microsoft.Extensions.Configuration.Abstractions" Version="8.0.0" />
+        <PackageReference Include="Microsoft.Extensions.DependencyInjection.Abstractions" Version="8.0.0" />
         <PackageReference Include="Scrutor" Version="4.2.2" />
         <PackageReference Include="Swashbuckle.AspNetCore" Version="6.5.0" />
-        <PackageReference Include="MediatR" Version="12.1.1" />
+        <PackageReference Include="MediatR" Version="12.2.0" />
         <PackageReference Include="AspNetCore.HealthChecks.Rabbitmq" Version="7.0.0" />
         <PackageReference Include="AspNetCore.HealthChecks.SqlServer" Version="7.0.0" />
         <PackageReference Include="AspNetCore.HealthChecks.UI" Version="7.0.2" />
         <PackageReference Include="AspNetCore.HealthChecks.UI.Client" Version="7.1.0" />
         <PackageReference Include="AspNetCore.HealthChecks.UI.InMemory.Storage" Version="7.0.0" />
         <PackageReference Include="FluentValidation.AspNetCore" Version="11.3.0" />
-        <PackageReference Include="Microsoft.AspNetCore.Mvc.NewtonsoftJson" Version="7.0.13" />
+        <PackageReference Include="Microsoft.AspNetCore.Mvc.NewtonsoftJson" Version="8.0.0" />
         <PackageReference Include="Microsoft.AspNetCore.OData" Version="8.2.3" />
     </ItemGroup>
 
