@@ -1,13 +1,12 @@
 ﻿using System;
 using System.Collections.Generic;
 using Wally.Lib.DDD.Abstractions.DomainEvents;
-using Wally.Lib.DDD.Abstractions.Validators;
 
 namespace Wally.RomMaster.HashService.Domain.Abstractions;
 
-public abstract class Entity<TEntity, TKey>
+public class Entity<TEntity, TKey>
 	where TEntity : Entity<TEntity, TKey>
-	where TKey : notnull, IComparable<TKey>, IEquatable<TKey>, IStronglyTypedId<TKey, Guid>, new()
+	where TKey : notnull, IComparable<TKey>, IEquatable<TKey>, new()
 {
 	private readonly List<DomainEvent> _domainEvents = new();
 
@@ -36,13 +35,5 @@ public abstract class Entity<TEntity, TKey>
 	public void RemoveDomainEvent(DomainEvent domainEvent)
 	{
 		_domainEvents.Remove(domainEvent);
-	}
-
-	protected static void CheckRule(IBusinessRule rule)
-	{
-		if (rule.IsBroken())
-		{
-			throw new NotImplementedException();
-		}
 	}
 }
