@@ -19,85 +19,10 @@ public static class HealthChecksExtensions
 	// Adds Readiness
 	public static IServiceCollection AddHealthChecks(this IServiceCollection services, AppSettings settings)
 	{
-		var healthChecksBuilder = services.AddHealthChecks()
-			.AddVersionHealthCheck();
-
-		switch (settings.Database.ProviderType)
-		{
-			case DatabaseProviderType.None:
-				break;
-			case DatabaseProviderType.MySql:
-				healthChecksBuilder.AddMySql(
-					new MySqlHealthCheckOptions { ConnectionString = settings.ConnectionStrings.Database, },
-					"DB",
-					HealthStatus.Degraded,
-					new[] { "DB", "Database", nameof(DatabaseProviderType.MySql), });
-				break;
-			case DatabaseProviderType.PostgreSQL:
-				healthChecksBuilder.AddNpgSql(
-					settings.ConnectionStrings.Database,
-					name: "DB",
-					failureStatus: HealthStatus.Degraded,
-					tags: new[] { "DB", "Database", nameof(DatabaseProviderType.PostgreSQL), });
-				break;
-			case DatabaseProviderType.SQLite:
-				healthChecksBuilder.AddSqlite(
-					settings.ConnectionStrings.Database,
-					name: "DB",
-					failureStatus: HealthStatus.Degraded,
-					tags: new[] { "DB", "Database", nameof(DatabaseProviderType.SQLite), });
-				break;
-			case DatabaseProviderType.SqlServer:
-				healthChecksBuilder.AddSqlServer(
-					settings.ConnectionStrings.Database,
-					name: "DB",
-					failureStatus: HealthStatus.Degraded,
-					tags: new[] { "DB", "Database", nameof(DatabaseProviderType.SqlServer), });
-				break;
-			default:
-				throw new ArgumentOutOfRangeException(
-					nameof(settings.Database.ProviderType),
-					"Unknown Database Provider Type");
-		}
-
-		switch (settings.MessageBroker)
-		{
-			case MessageBrokerType.None:
-				break;
-			case MessageBrokerType.AzureServiceBus:
-				// TODO: Add AzureServiceBus HealthCheck
-				// healthChecksBuilder.AddAzureServiceBusQueue()
-				// ...
-
-				break;
-			case MessageBrokerType.Kafka:
-				healthChecksBuilder.AddKafka(
-					new KafkaHealthCheckOptions
-					{
-						Configuration = new ProducerConfig(
-							new ClientConfig { BootstrapServers = settings.ConnectionStrings.ServiceBus, }),
-					},
-					"MQ",
-					HealthStatus.Degraded,
-					new[] { "MQ", "Messaging", nameof(MessageBrokerType.Kafka), });
-				break;
-			case MessageBrokerType.RabbitMQ:
-				healthChecksBuilder.AddRabbitMQ(
-					new Uri(settings.ConnectionStrings.ServiceBus),
-					name: "MQ",
-					failureStatus: HealthStatus.Degraded,
-					tags: new[] { "MQ", "Messaging", nameof(MessageBrokerType.RabbitMQ), });
-				break;
-			default:
-				throw new ArgumentOutOfRangeException(
-					nameof(settings.MessageBroker),
-					$"Unknown Message Broker: '{settings.MessageBroker}'");
-		}
-
-		/*
-		services.AddHealthChecksUI() // TODO: Consider only for ApiGateway
-			.AddInMemoryStorage();
-		*/
+		services.AddHealthChecks()
+			.WithVersion()
+			.WithDatabase(settings)
+			.WithMessageBroker(settings);
 
 		return services;
 	}
@@ -108,11 +33,9 @@ public static class HealthChecksExtensions
 			"/healthChecks",
 			new HealthCheckOptions
 			{
-				Predicate = _ => true, ResponseWriter = UIResponseWriter.WriteHealthCheckUIResponse,
+				Predicate = _ => true,
+				ResponseWriter = UIResponseWriter.WriteHealthCheckUIResponse,
 			});
-		/*
-		app.UseHealthChecksUI(); // TODO: Consider only for ApiGateway
-		*/
 
 		app.UseEndpoints(
 			endpoints =>
@@ -131,6 +54,140 @@ public static class HealthChecksExtensions
 		return app;
 	}
 
+	private static IHealthChecksBuilder WithVersion(this IHealthChecksBuilder healthChecksBuilder)
+	{
+		return healthChecksBuilder.AddVersionHealthCheck();
+	}
+
+	private static IHealthChecksBuilder WithDatabase(this IHealthChecksBuilder healthChecksBuilder,
+		AppSettings settings)
+	{
+		switch (settings.Database.ProviderType)
+		{
+			case DatabaseProviderType.None:
+				break;
+			case DatabaseProviderType.InMemory:
+				break;
+			case DatabaseProviderType.MySql:
+				healthChecksBuilder.AddMySql(
+					new MySqlHealthCheckOptions
+					{
+						ConnectionString = settings.ConnectionStrings.Database,
+					},
+					"DB",
+					HealthStatus.Degraded,
+					new[]
+					{
+						"DB", "Database", nameof(DatabaseProviderType.MySql),
+					});
+				break;
+			case DatabaseProviderType.PostgreSQL:
+				healthChecksBuilder.AddNpgSql(
+					settings.ConnectionStrings.Database,
+					name: "DB",
+					failureStatus: HealthStatus.Degraded,
+					tags: new[]
+					{
+						"DB", "Database", nameof(DatabaseProviderType.PostgreSQL),
+					});
+				break;
+			case DatabaseProviderType.SQLite:
+				healthChecksBuilder.AddSqlite(
+					settings.ConnectionStrings.Database,
+					name: "DB",
+					failureStatus: HealthStatus.Degraded,
+					tags: new[]
+					{
+						"DB", "Database", nameof(DatabaseProviderType.SQLite),
+					});
+				break;
+			case DatabaseProviderType.SqlServer:
+				healthChecksBuilder.AddSqlServer(
+					settings.ConnectionStrings.Database,
+					name: "DB",
+					failureStatus: HealthStatus.Degraded,
+					tags: new[]
+					{
+						"DB", "Database", nameof(DatabaseProviderType.SqlServer),
+					});
+				break;
+			default:
+				throw new ArgumentOutOfRangeException(nameof(settings.Database.ProviderType),
+					"Unknown Database Provider Type");
+		}
+
+		return healthChecksBuilder;
+	}
+
+	private static IHealthChecksBuilder WithMessageBroker(this IHealthChecksBuilder healthChecksBuilder,
+		AppSettings settings)
+	{
+		switch (settings.MessageBroker)
+		{
+			case MessageBrokerType.None:
+				break;
+			case MessageBrokerType.AzureServiceBus:
+				healthChecksBuilder.WithAzureServiceBus();
+				break;
+			case MessageBrokerType.Kafka:
+				healthChecksBuilder.WithKafka(settings);
+				break;
+			case MessageBrokerType.RabbitMQ:
+				healthChecksBuilder.WithRabbitMQ(settings);
+				break;
+			default:
+				throw new ArgumentOutOfRangeException(nameof(settings.MessageBroker),
+					$"Unknown Message Broker: '{settings.MessageBroker}'");
+		}
+
+		return healthChecksBuilder;
+	}
+
+	private static IHealthChecksBuilder WithAzureServiceBus(this IHealthChecksBuilder healthChecksBuilder)
+	{
+		// TODO: Add AzureServiceBus HealthCheck
+		// healthChecksBuilder.AddAzureServiceBusQueue()
+		// ...
+
+		return healthChecksBuilder;
+	}
+
+	private static IHealthChecksBuilder WithRabbitMQ(this IHealthChecksBuilder healthChecksBuilder,
+		AppSettings settings)
+	{
+		healthChecksBuilder.AddRabbitMQ(
+			new Uri(settings.ConnectionStrings.ServiceBus),
+			name: "MQ",
+			failureStatus: HealthStatus.Degraded,
+			tags: new[]
+			{
+				"MQ", "Messaging", nameof(MessageBrokerType.RabbitMQ),
+			});
+
+		return healthChecksBuilder;
+	}
+
+	private static IHealthChecksBuilder WithKafka(this IHealthChecksBuilder healthChecksBuilder, AppSettings settings)
+	{
+		healthChecksBuilder.AddKafka(
+			new KafkaHealthCheckOptions
+			{
+				Configuration = new ProducerConfig(
+					new ClientConfig
+					{
+						BootstrapServers = settings.ConnectionStrings.ServiceBus,
+					}),
+			},
+			"MQ",
+			HealthStatus.Degraded,
+			new[]
+			{
+				"MQ", "Messaging", nameof(MessageBrokerType.Kafka),
+			});
+
+		return healthChecksBuilder;
+	}
+
 	private static IHealthChecksBuilder AddVersionHealthCheck(this IHealthChecksBuilder builder)
 	{
 		builder.AddCheck<VersionHealthCheck>("VER", tags: new[] { "VER", "Version", });
@@ -138,18 +195,13 @@ public static class HealthChecksExtensions
 		return builder;
 	}
 
-	private class VersionHealthCheck : IHealthCheck
+	private sealed class VersionHealthCheck : IHealthCheck
 	{
-		private readonly string? _version;
+		private readonly string? _version = typeof(VersionHealthCheck).Assembly.GetName()
+			.Version?.ToString();
 
-		public VersionHealthCheck()
-		{
-			_version = GetType()
-				.Assembly.GetName()
-				.Version?.ToString();
-		}
-
-		public Task<HealthCheckResult> CheckHealthAsync(HealthCheckContext context, CancellationToken cancellationToken)
+		public Task<HealthCheckResult> CheckHealthAsync(HealthCheckContext context,
+			CancellationToken cancellationToken = default)
 		{
 			return Task.FromResult(HealthCheckResult.Healthy(_version));
 		}
