@@ -36,7 +36,10 @@ internal class ODataQueryOptionsOperationFilter : IOperationFilter
 				In = ParameterLocation.Query,
 				Description = "Filter the results",
 				Required = false,
-				Schema = new OpenApiSchema { Type = "string", },
+				Schema = new OpenApiSchema
+				{
+					Type = "string",
+				},
 			});
 
 		operation.Parameters.Add(
@@ -46,7 +49,10 @@ internal class ODataQueryOptionsOperationFilter : IOperationFilter
 				In = ParameterLocation.Query,
 				Description = "Search term",
 				Required = false,
-				Schema = new OpenApiSchema { Type = "string", },
+				Schema = new OpenApiSchema
+				{
+					Type = "string",
+				},
 			});
 
 		operation.Parameters.Add(
@@ -56,7 +62,10 @@ internal class ODataQueryOptionsOperationFilter : IOperationFilter
 				In = ParameterLocation.Query,
 				Description = "Order the results",
 				Required = false,
-				Schema = new OpenApiSchema { Type = "string", },
+				Schema = new OpenApiSchema
+				{
+					Type = "string",
+				},
 			});
 
 		operation.Parameters.Add(
@@ -66,7 +75,11 @@ internal class ODataQueryOptionsOperationFilter : IOperationFilter
 				In = ParameterLocation.Query,
 				Description = "Select the properties to be returned in the response",
 				Required = false,
-				Schema = new OpenApiSchema { Type = "string", Deprecated = true, },
+				Schema = new OpenApiSchema
+				{
+					Type = "string",
+					Deprecated = true,
+				},
 			});
 
 		operation.Parameters.Add(
@@ -76,7 +89,12 @@ internal class ODataQueryOptionsOperationFilter : IOperationFilter
 				In = ParameterLocation.Query,
 				Description = "Limit the number of results returned",
 				Required = false,
-				Schema = new OpenApiSchema { Type = "integer", Format = "int32", Minimum = 0, },
+				Schema = new OpenApiSchema
+				{
+					Type = "integer",
+					Format = "int32",
+					Minimum = 0,
+				},
 			});
 
 		operation.Parameters.Add(
@@ -86,7 +104,12 @@ internal class ODataQueryOptionsOperationFilter : IOperationFilter
 				In = ParameterLocation.Query,
 				Description = "Skip the specified number of results",
 				Required = false,
-				Schema = new OpenApiSchema { Type = "integer", Format = "int32", Minimum = 0, },
+				Schema = new OpenApiSchema
+				{
+					Type = "integer",
+					Format = "int32",
+					Minimum = 0,
+				},
 			});
 	}
 }
