@@ -5,9 +5,9 @@ using FluentAssertions;
 using FluentAssertions.Execution;
 using FluentAssertions.Types;
 using Microsoft.AspNetCore.Mvc;
+using Wally.RomMaster.HashService.Tests.ConventionTests.Helpers;
 using Wally.Lib.DDD.Abstractions.Commands;
 using Wally.Lib.DDD.Abstractions.Queries;
-using Wally.RomMaster.HashService.Tests.ConventionTests.Helpers;
 using Xunit;
 
 namespace Wally.RomMaster.HashService.Tests.ConventionTests;
