@@ -9,13 +9,13 @@ public class BusStub : IBus
 {
 	public ConnectHandle ConnectPublishObserver(IPublishObserver observer)
 	{
-		throw new NotImplementedException();
+		throw new NotSupportedException();
 	}
 
 	public Task<ISendEndpoint> GetPublishSendEndpoint<T>()
 		where T : class
 	{
-		throw new NotImplementedException();
+		throw new NotSupportedException();
 	}
 
 	public Task Publish<T>(T message, CancellationToken cancellationToken = new())
@@ -27,28 +27,28 @@ public class BusStub : IBus
 	public Task Publish<T>(T message, IPipe<PublishContext<T>> publishPipe, CancellationToken cancellationToken = new())
 		where T : class
 	{
-		throw new NotImplementedException();
+		throw new NotSupportedException();
 	}
 
 	public Task Publish<T>(T message, IPipe<PublishContext> publishPipe, CancellationToken cancellationToken = new())
 		where T : class
 	{
-		throw new NotImplementedException();
+		throw new NotSupportedException();
 	}
 
 	public Task Publish(object message, CancellationToken cancellationToken = new())
 	{
-		throw new NotImplementedException();
+		throw new NotSupportedException();
 	}
 
 	public Task Publish(object message, IPipe<PublishContext> publishPipe, CancellationToken cancellationToken = new())
 	{
-		throw new NotImplementedException();
+		throw new NotSupportedException();
 	}
 
 	public Task Publish(object message, Type messageType, CancellationToken cancellationToken = new())
 	{
-		throw new NotImplementedException();
+		throw new NotSupportedException();
 	}
 
 	public Task Publish(
@@ -57,13 +57,13 @@ public class BusStub : IBus
 		IPipe<PublishContext> publishPipe,
 		CancellationToken cancellationToken = new())
 	{
-		throw new NotImplementedException();
+		throw new NotSupportedException();
 	}
 
 	public Task Publish<T>(object values, CancellationToken cancellationToken = new())
 		where T : class
 	{
-		throw new NotImplementedException();
+		throw new NotSupportedException();
 	}
 
 	public Task Publish<T>(
@@ -72,7 +72,7 @@ public class BusStub : IBus
 		CancellationToken cancellationToken = new())
 		where T : class
 	{
-		throw new NotImplementedException();
+		throw new NotSupportedException();
 	}
 
 	public Task Publish<T>(
@@ -81,61 +81,61 @@ public class BusStub : IBus
 		CancellationToken cancellationToken = new())
 		where T : class
 	{
-		throw new NotImplementedException();
+		throw new NotSupportedException();
 	}
 
 	public ConnectHandle ConnectSendObserver(ISendObserver observer)
 	{
-		throw new NotImplementedException();
+		throw new NotSupportedException();
 	}
 
 	public Task<ISendEndpoint> GetSendEndpoint(Uri address)
 	{
-		throw new NotImplementedException();
+		throw new NotSupportedException();
 	}
 
 	public ConnectHandle ConnectConsumePipe<T>(IPipe<ConsumeContext<T>> pipe)
 		where T : class
 	{
-		throw new NotImplementedException();
+		throw new NotSupportedException();
 	}
 
 	public ConnectHandle ConnectConsumePipe<T>(IPipe<ConsumeContext<T>> pipe, ConnectPipeOptions options)
 		where T : class
 	{
-		throw new NotImplementedException();
+		throw new NotSupportedException();
 	}
 
 	public ConnectHandle ConnectRequestPipe<T>(Guid requestId, IPipe<ConsumeContext<T>> pipe)
 		where T : class
 	{
-		throw new NotImplementedException();
+		throw new NotSupportedException();
 	}
 
 	public ConnectHandle ConnectConsumeMessageObserver<T>(IConsumeMessageObserver<T> observer)
 		where T : class
 	{
-		throw new NotImplementedException();
+		throw new NotSupportedException();
 	}
 
 	public ConnectHandle ConnectConsumeObserver(IConsumeObserver observer)
 	{
-		throw new NotImplementedException();
+		throw new NotSupportedException();
 	}
 
 	public ConnectHandle ConnectReceiveObserver(IReceiveObserver observer)
 	{
-		throw new NotImplementedException();
+		throw new NotSupportedException();
 	}
 
 	public ConnectHandle ConnectReceiveEndpointObserver(IReceiveEndpointObserver observer)
 	{
-		throw new NotImplementedException();
+		throw new NotSupportedException();
 	}
 
 	public ConnectHandle ConnectEndpointConfigurationObserver(IEndpointConfigurationObserver observer)
 	{
-		throw new NotImplementedException();
+		throw new NotSupportedException();
 	}
 
 	public HostReceiveEndpointHandle ConnectReceiveEndpoint(
@@ -143,19 +143,19 @@ public class BusStub : IBus
 		IEndpointNameFormatter? endpointNameFormatter = null,
 		Action<IReceiveEndpointConfigurator>? configureEndpoint = null)
 	{
-		throw new NotImplementedException();
+		throw new NotSupportedException();
 	}
 
 	public HostReceiveEndpointHandle ConnectReceiveEndpoint(
 		string queueName,
 		Action<IReceiveEndpointConfigurator>? configureEndpoint = null)
 	{
-		throw new NotImplementedException();
+		throw new NotSupportedException();
 	}
 
 	public void Probe(ProbeContext context)
 	{
-		throw new NotImplementedException();
+		throw new NotSupportedException();
 	}
 
 	public Uri Address { get; } = new("http://localhost");
